pyproject.toml
[project]
name = "fvctools"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Boris Resnick", email = "boris@flyver.city" }
]
dependencies = [
    "pyparsing>=3.2.0",
    "toolz>=1.0.0",
    "jsonschema>=4.23.0",
    "pygeodesy>=24.11.11",
    "scipy>=1.14.1",
    "pynmea2>=1.19.0",
    "python-dateutil>=2.9.0.post0",
    "gitpython>=3.1.43",
    "click>=8.1.7",
    "simplekml>=1.3.6",
    "pygments>=2.18.0",
    "boto3>=1.35.76",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
fvc = { cmd = ['python', 'src/fvc/tools/cli.py'] }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fvc"]
